+++ Gruntfile.js
@@ -1,23 +1,25 @@
 module.exports = function(grunt){
-  var path          = require('path');
+  var path  = require('path');
+  var babel = require('rollup-plugin-babel');
+  var resolve = require('rollup-plugin-node-resolve');
+
   require('time-grunt')(grunt);
 
-  var projectName = 'KRAIT';
+  var projectName = 'Krait';
   var projectNameLC = projectName.toLowerCase();
 
   var port      = 3008;
   var host      = 'localhost';
 
   var srcDir          = 'src/';
-  var compiledSrcDir  = srcDir + 'build/';
+  var compiledSrcDir  = 'build/';
+  var compiledES5Dir  = compiledSrcDir + 'es5/';
+  var compiledES6Dir  = compiledSrcDir + 'es6/';
   var distDir         = 'dist/';
   var webDir          = 'website/';
   var publicDir       = webDir + 'public/';
   var nodeDir         = 'node_modules/';
-  var bowerDir        = 'bower_components/';
   var docDir          = 'doc/';
-  var zipDir          = 'zip/';
-
 
   var banner    = '/** MIT License\n' +
     '* \n' +
@@ -44,41 +46,34 @@ module.exports = function(grunt){
     '* http://' + projectNameLC + 'js.lcluber.com\n' +
     '*/\n';
 
-  //i18n configuration for static website translations
-  var i18n = require('i18next');
-  i18n.init({
-    lng: 'en',
-    resources: {
-      en: {
-        translation: grunt.file.readJSON( webDir + 'locales/en/translation.json' )
-      }
-    }
-  });
+
   grunt.option('stack', true);
   // Project configuration.
   grunt.initConfig({
     pkg: grunt.file.readJSON('package.json'),
     clean: {
       lib:{
+        dot: true,
         src: [  distDir + '*',
-                compiledSrcDir + '*'
+                compiledES5Dir + '*',
+                compiledES6Dir + '*'
               ]
       },
       doc:{
         src: [  docDir + '*'
               ]
       },
-      web:{
-        src: [  zipDir + '*',
-                webDir + 'static/*',
-                webDir + 'sass/build/*',
-                webDir + 'js/build/*'
+      websass:{
+        src: [  webDir + 'sass/build/*',
+                publicDir + 'css/*'
         ]
       },
-      public: {
-        src: [  publicDir + 'js/*',
-                publicDir + 'css/*',
-                publicDir + 'fonts/*'
+      webjs:{
+        src: [  publicDir + 'js/*'
+        ]
+      },
+      webmisc: {
+        src: [  publicDir + 'fonts/*'
         ]
       }
     },
@@ -87,17 +82,18 @@ module.exports = function(grunt){
   			options: {
   				out: docDir,
   				target: 'es6',
-          name: projectName + '.js - Documentation'
+          name: projectName + '.js - Documentation',
+          includeDeclarations: false,
+          hideGenerator: true
   			},
   			src: [srcDir + 'ts/*.ts']
   		}
   	},
     jshint: {
       options: {
-        jshintrc: 'config/.jshintrc'
+        //jshintrc: 'config/.jshintrc'
       },
-      lib: [ 'Gruntfile.js', srcDir + '**/*.js'],
-      web: [ webDir + 'js/*.js'],
+      web: [ webDir + 'js/*.js']
     },
     sass: {
       dist: {
@@ -133,108 +129,92 @@ module.exports = function(grunt){
         }]
       }
     },
-    // jsdoc: {
-    //   dist : {
-    //     src: distDir + projectNameLC + '.js',
-    //     config: 'config/jsdoc-conf.json'
+    // htmlmin: {
+    //   static: {
+    //     options: {
+    //       removeComments: true,
+    //       collapseWhitespace: true
+    //     },
+    //     expand: true,
+    //     cwd: webDir + 'static',
+    //     src: ['**/*.htm'],
+    //     dest: webDir + 'static/'
     //   }
     // },
-    pug: {
-      compile: {
-        options: {
-          namespace   : 'JST',
-          separator   : '\n\n',
-          amd         : false,
-          client      : false,
-          pretty      : true,
-          self        : false,
-          debug       : false,
-          compileDebug: true,
-          data: function() {
-            return {
-              t: i18n.t.bind(i18n)
-            };
-          }
-        },
-        files: [ {
-          cwd: webDir + 'views',
-          src: ['**/*.pug', '!**/_*.pug'],
-          dest: webDir + 'static',
-          expand: true,
-          ext: '.htm'
-        } ]
-      }
-    },
-    htmlmin: {
-      static: {
-        options: {
-          removeComments: true,
-          collapseWhitespace: true
-        },
-        expand: true,
-        cwd: webDir + 'static',
-        src: ['**/*.htm'],
-        dest: webDir + 'static/'
-      }
-    },
-    bower_concat: {
-      all: {
-        dest: {
-          'js': webDir + 'js/build/bower.js'
-          //'css': 'build/_bower.css'
-        },
-        exclude: [
-        ],
-        dependencies: {
-        },
-        bowerOptions: {
-          relative: false
-        }
-      }
-    },
-    tslint: {
-      options: {
-        configuration: 'config/tslint.json',
-        force: false
-      },
-      lib: {
-        files: [{
-          expand: true,
-          cwd: srcDir,
-          src: [ srcDir + '**/*.ts' ]
-        }]
-      }
-    },
+    // tslint: {
+    //   options: {
+    //     configuration: 'config/tslint.json',
+    //     force: false
+    //   },
+    //   lib: {
+    //     files: [{
+    //       expand: true,
+    //       cwd: srcDir,
+    //       src: [ srcDir + '**/*.ts' ]
+    //     }]
+    //   }
+    // },
     ts: {
-      tsconfig: 'config/tsconfig.json',
-      lib : {
-        outDir: compiledSrcDir,
-        options: {
-          rootDir: srcDir + 'ts/',
-          declaration: true
-        },
+      options: {
+        fast: 'never'
+      },
+      es6: {
+        tsconfig: './tsconfig.json',
+        src: [ srcDir + 'ts/**/*.ts', '!node_modules/**/*.ts' ]
+      },
+      es5: {
+        tsconfig: './tsconfig.es5.json',
         src: [ srcDir + 'ts/**/*.ts', '!node_modules/**/*.ts' ]
       }
     },
     rollup: {
-      options: {
-        format:'umd',
-        moduleName: projectName,
-        external: [
-          path.resolve( './bower_components/Weejs/dist/wee.js' ),
-          path.resolve( './bower_components/Mouettejs/dist/mouette.js' )
-        ],
-        banner: banner
-      },
-      bundle:{
+      es6: {
+        options: {
+          format:'es',
+          // moduleName: projectName,
+          banner: banner,
+          // sourceMap: 'inline'
+          plugins: [
+            resolve({
+            //   //exclude: './node_modules/**'
+            })
+          ],
+          external: [
+            'mouettejs',
+            'weejs'
+          ]
+        },
         files: [ {
-          src : compiledSrcDir + projectNameLC + '.js',
+          src : compiledES6Dir + projectNameLC + '.js',
           dest : distDir + projectNameLC + '.js'
+        } ]
+      },
+      iife: {
+        options: {
+          format:'iife',
+          moduleName: projectName,
+          banner: banner,
+          plugins: [
+            babel({
+            //   //exclude: './node_modules/**'
+            }),
+            resolve({
+              //   //exclude: './node_modules/**'
+            })
+          ]
+          // sourceMap: 'inline'
+          // external: [
+
+          // ]
+        },
+        files: [ {
+          src : compiledES5Dir + projectNameLC + '.js',
+          dest : distDir + projectNameLC + '.iife.js'
         } ]
       }
     },
     uglify: {
-      libmin: {
+      libIife: {
         options: {
           sourceMap: false,
           sourceMapName: srcDir + 'sourcemap.map',
@@ -255,48 +235,14 @@ module.exports = function(grunt){
             hoist_funs:true,
             if_return:true,
             join_vars:true,
-            cascade:true,
             warnings: true,
             drop_console: false,
             keep_fargs: false,
             keep_fnames: false
           }
         },
-        src: distDir + projectNameLC + '.js',
-        dest: distDir + projectNameLC + '.min.js'
-      },
-      bower: {
-        options: {
-          sourceMap: false,
-          sourceMapName: srcDir + 'sourcemap.map',
-          mangle: {
-            reserved: []
-          },
-          banner: '',
-          compress: {
-            sequences: true,
-            properties: true,
-            dead_code: true,
-            unsafe: false,
-            conditionals:true,
-            comparisons:true,
-            booleans:true,
-            loops:true,
-            unused: true,
-            hoist_funs:true,
-            if_return:true,
-            join_vars:true,
-            cascade:true,
-            warnings: true,
-            drop_console: false,
-            keep_fargs: false,
-            keep_fnames: false
-          }
-        },
-        files: [{
-          src: webDir + 'js/build/*.js',
-          dest : webDir + 'js/build/bower.min.js'
-        }]
+        src: distDir + projectNameLC + '.iife.js',
+        dest: distDir + projectNameLC + '.iife.min.js'
       },
       web: {
         options: {
@@ -353,8 +299,7 @@ module.exports = function(grunt){
         },
         src: [  nodeDir   + 'jquery/dist/jquery.min.js',
                 nodeDir   + 'bootstrap/dist/js/bootstrap.min.js',
-                webDir + 'js/build/bower.min.js',
-                distDir   + projectNameLC + '.min.js',
+                distDir   + projectNameLC + '.iife.min.js',
                 publicDir + 'js/main.min.js'
             ],
         dest: publicDir + 'js/main.min.js'
@@ -367,7 +312,7 @@ module.exports = function(grunt){
         },
         src: [nodeDir + 'font-awesome/css/font-awesome.min.css',
               nodeDir + 'bootstrap/dist/css/bootstrap.min.css',
-              bowerDir + 'Mouettejs/dist/mouette.css',
+              nodeDir + 'mouettejs/dist/mouette.css',
               publicDir + 'css/style.min.css'
             ],
         dest: publicDir + 'css/style.min.css'
@@ -389,16 +334,10 @@ module.exports = function(grunt){
     copy: {
       mouette:{
         expand: true,
-        cwd: bowerDir + 'mouettejs/dist/',
+        cwd: nodeDir + 'mouettejs/dist/',
         src: ['*.htm'],
         dest: webDir + 'views/',
         filter: 'isFile'
-      }
-    },
-    symlink: {
-      options: {
-        overwrite: false,
-        force: false
       },
       fonts:{
         expand: true,
@@ -413,36 +352,6 @@ module.exports = function(grunt){
         src: ['fonts/**/*'],
         dest: publicDir,
         filter: 'isFile'
-      },
-      public: {
-        expand: true,
-        cwd: publicDir,
-        src: ['**/*'],
-        dest: webDir + 'static/public/'
-      },
-      doc: {
-        expand: true,
-        cwd: docDir,
-        src: ['**/*'],
-        dest: webDir + 'static/' + docDir
-      }
-    },
-    compress: {
-      main: {
-        options: {
-          archive: zipDir + projectNameLC + 'js.zip'
-        },
-        files: [
-          {expand: true, cwd: webDir + 'static/', src: '**', dest: '/'},
-          {expand: true, cwd: publicDir, src: '**', dest: '/public'},
-          {src: [ distDir + '**',
-                  docDir + '**',
-                  'LICENCE.md',
-                  'README.md',
-                  'RELEASE_NOTES.md'
-                ],
-                dest: '/', filter: 'isFile'}
-        ]
       }
     },
     nodemon: {
@@ -463,19 +372,19 @@ module.exports = function(grunt){
     },
     watch: {
       lib: {
-        files: srcDir + 'ts/**/*.ts',
-        tasks: ['dist'],
+        files: [ srcDir + 'ts/**/*.ts', '!' + srcDir + 'ts/build/*'],
+        tasks: ['lib', 'webjs']
       },
       webpug:{
         files: webDir + 'views/**/*.pug'
       },
       webjs: {
         files: webDir + 'js/**/*.js',
-        tasks: ['website'],
+        tasks: ['webjs']
       },
       webcss: {
-        files: webDir + 'sass/**/*.scss',
-        tasks: ['website'],
+        files: [webDir + 'sass/**/*.scss', '!' + webDir + 'sass/build/*'],
+        tasks: ['websass']
       },
       options: {
         interrupt: true,
@@ -493,7 +402,6 @@ module.exports = function(grunt){
     }
   });
 
-  grunt.loadNpmTasks( 'grunt-bower-concat' );
   grunt.loadNpmTasks( 'grunt-contrib-copy' );
   grunt.loadNpmTasks( 'grunt-contrib-clean' );
   grunt.loadNpmTasks( 'grunt-contrib-jshint' );
@@ -501,14 +409,10 @@ module.exports = function(grunt){
   grunt.loadNpmTasks( 'grunt-contrib-csslint' );
   grunt.loadNpmTasks( 'grunt-contrib-cssmin' );
   grunt.loadNpmTasks( 'grunt-contrib-concat' );
-  grunt.loadNpmTasks( 'grunt-contrib-pug' );
   grunt.loadNpmTasks( 'grunt-contrib-sass' );
   grunt.loadNpmTasks( 'grunt-contrib-htmlmin' );
-  grunt.loadNpmTasks( 'grunt-contrib-symlink' );
-  grunt.loadNpmTasks( 'grunt-contrib-compress' );
   grunt.loadNpmTasks( 'grunt-contrib-watch' );
   grunt.loadNpmTasks( 'grunt-strip-code' );
-  grunt.loadNpmTasks( 'grunt-jsdoc' );
   grunt.loadNpmTasks( 'grunt-concurrent' );
   grunt.loadNpmTasks( 'grunt-nodemon' );
   grunt.loadNpmTasks( 'grunt-open' );
@@ -519,14 +423,19 @@ module.exports = function(grunt){
 
   grunt.registerTask( 'lib',
                       'build the library in the dist/ folder',
-                      [ 'tslint:lib',
+                      [ //'tslint:lib',
                         'clean:lib',
-                        'ts:lib',
-                        'rollup',
-                        'uglify:libmin',
+                        //lib es6
+                        'ts:es6',
+                        'rollup:es6',
+                        //lib es5
+                        'ts:es5',
+                        'rollup:iife',
+                        'uglify:libIife',
+                        //declaration
                         'concat:declaration',
                         'strip_code:declaration'
-                        //'jsdoc'
+                        //'replace:declaration'
                       ]
                     );
 
@@ -537,48 +446,48 @@ module.exports = function(grunt){
                       ]
                     );
 
-  // grunt.registerTask( 'static',
-  //                     'build static version of the website into /website/static',
-  //                     [ 'pug',
-  //                       'htmlmin',
-  //                       'symlink:fonts', 'symlink:fontAwesome', 'symlink:public', 'symlink:doc'
-  //                     ]
-  //                   );
-
-  // grunt.registerTask( 'zip',
-  //                     'create the zip package',
-  //                     ['compress']
-  //                   );
-
   grunt.registerTask( 'serve',
                       'launch server, open website and watch for changes',
                       [ 'concurrent' ]
                     );
 
-  grunt.registerTask( 'website',
-                      'build the website in the website/ folder',
-                      [ 'jshint:web',
-                        'clean:public', 'clean:web',
-                        //js
-                          'bower_concat',
-                          'uglify:bower',
-                          'uglify:web',
-                          'concat:webjs',
-                        //css
-                          'sass',
-                          'cssmin',
-                          'symlink:fonts', 'symlink:fontAwesome',
-                          'copy:mouette',
-                          'concat:webcss',
-                        //static
-                          'pug',
-                          'htmlmin',
-                          'symlink:public',
-                          'compress'
-                      ]
+  grunt.registerTask( 'websass',
+                      'Compile website css',
+                      [ 'clean:websass',
+                        'sass',
+                        'cssmin',
+                        'copy:mouette',
+                        'concat:webcss'
+                       ]
                     );
 
-  grunt.registerTask( 'dist',
+  grunt.registerTask( 'webjs',
+                      'Compile website js',
+                      [ 'jshint:web',
+                        'clean:webjs',
+                        'uglify:web',
+                        'concat:webjs'
+                       ]
+                    );
+
+  grunt.registerTask( 'webmisc',
+                      'Compile website misc',
+                      [ 'clean:webmisc',
+                        'copy:fonts',
+                        'copy:fontAwesome'
+                       ]
+                    );
+
+  grunt.registerTask( 'website',
+                      'build the website in the website/ folder',
+                      function() {
+                        grunt.task.run('webjs');
+                        grunt.task.run('websass');
+                        grunt.task.run('webmisc');
+                      }
+                    );
+
+  grunt.registerTask( 'build',
                       'build library and website',
                       function() {
                         //build lib
@@ -587,14 +496,14 @@ module.exports = function(grunt){
                         grunt.task.run('website');
                         //build documentation
                         grunt.task.run('doc');
+                        // launch server and watch for changes
+                        grunt.task.run('serve');
                       }
                     );
 
   grunt.registerTask( 'default',
                       'build library, website, launch server, open website and watch for changes.',
                       function() {
-                        //build library and website
-                        grunt.task.run('dist');
                         // launch server and watch for changes
                         grunt.task.run('serve');
                       }